Comentarios

+++ Activities/Controllers/RowsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Activities.Models;
 using Activities.Helpers;
+using Microsoft.AspNetCore.Http;
 
 namespace Activities.Controllers
 {
@@ -22,12 +23,18 @@ namespace Activities.Controllers
         // GET: RowEntities
         public async Task<IActionResult> Index(int idActivity)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             return View(await _context.Row.Where(x => x.ActivityId == idActivity && x.IsActive == true).ToListAsync());
         }
 
         // GET: RowEntities/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             if (id == null)
             {
                 return NotFound();
@@ -46,6 +53,9 @@ namespace Activities.Controllers
         // GET: RowEntities/Create
         public IActionResult Create(int idActivity)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             ViewBag.IdActivity = idActivity;
             ViewData["ListRowStatus"] = new SelectList(_context.RowStatus.Where(x => x.IsActive).ToArray(), SelectProperty.Id, SelectProperty.Name);
 
@@ -59,9 +69,12 @@ namespace Activities.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ActivityId,UserId,RowStatusId,DateStop,DateStart,Name,Description,Id,DateRegistration,IsActive")] RowEntity rowEntity)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             rowEntity.IsActive = true;
             rowEntity.DateRegistration = DateTime.Now;
-
+            rowEntity.UserId = (int)HttpContext.Session.GetInt32(SessionUser.Id);
             if (ModelState.IsValid)
             {
                 _context.Add(rowEntity);
@@ -80,6 +93,9 @@ namespace Activities.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> ChangeRowStatus([Bind("ActivityId,UserId,RowStatusId,DateStop,DateStart,Name,Description,Id,DateRegistration,IsActive")] RowEntity rowEntity)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             RowEntity rowEntityOriginal;
 
             rowEntityOriginal = _context.Row.Where(x => x.Id == rowEntity.Id).FirstOrDefault();
@@ -92,6 +108,9 @@ namespace Activities.Controllers
         // GET: RowEntities/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             if (id == null)
             {
                 return NotFound();
@@ -112,6 +131,9 @@ namespace Activities.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("IdActivity,IdUser,IdRowStatus,DateStop,DateStart,Name,Description,Id,DateRegistration,IsActive")] RowEntity rowEntity)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             if (id != rowEntity.Id)
             {
                 return NotFound();
@@ -143,6 +165,9 @@ namespace Activities.Controllers
         // GET: RowEntities/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             if (id == null)
             {
                 return NotFound();
@@ -163,6 +188,9 @@ namespace Activities.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetInt32(SessionUser.Id) is null)
+                return RedirectToAction("Index", "Login");
+
             var rowEntity = await _context.Row.FindAsync(id);
             _context.Row.Remove(rowEntity);
             await _context.SaveChangesAsync();
